requirement.txt
streamlit
pymupdf
python-dotenv
apify-client